requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
requests
<<<<<<< HEAD
kagglehub
python-dotenv
kaggle
=======
python-dotenv
kaggle
notebook
ipykernel
nbconvert
>>>>>>> 1d092d5 (Add pipeline + orchestrator + tests + progress report; ignore data/results/secrets)
